requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
certifi==2023.7.22
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
coverage[toml]==7.2.7
    # via pytest-cov
cryptography==41.0.3
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.2.3
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-extensions
django-allauth==0.54.0
    # via -r requirements.in
django-environ==0.10.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-simple-history==3.3.0
    # via -r requirements.in
execnet==2.0.2
    # via pytest-xdist
factory-boy==3.3.0
    # via -r requirements.in
faker==19.2.0
    # via factory-boy
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.1
    # via pytest
pluggy==1.2.0
    # via pytest
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via django-allauth
pytest==7.4.0
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-socket
    #   pytest-xdist
pytest-cov==4.1.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
pytest-socket==0.6.0
    # via -r requirements.in
pytest-xdist==3.3.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
python3-openid==3.2.0
    # via django-allauth
pyupgrade==3.10.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
sentry-sdk==1.28.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tokenize-rt==5.2.0
    # via pyupgrade
urllib3==2.0.4
    # via
    #   requests
    #   sentry-sdk

# The following packages are considered to be unsafe in a requirements file:
# setuptools
